Update limiter and sedimentation output. Note bug in ns conservation limiter.
Updated limiter stats to write a summary to an output text file and save
full limiter stats data to the NetCDF output file. Added number of
sedimentation steps to the NetCDF output file.

Noted bug in ns conservatin limiter, where if nsubs and nsagg are zero
there is a division by zero and the scaling ratio for the limiter is
infinite. Commented out ns conservation limiter ouput to the NetCDF file
because infinite values cause writing output to crash. The bad limiter
values does not seem to effect results so far but need to investigate
further.