Add new mg2 module
This mg2.py file imports _mg2 and provides shadowed access to that
module. It also provides a tendency class that (slightly) improves the
interface to the micro_mg_tend() subroutine.

Much of the dense code that was in the first Jupyter Notebook has been
encapsulated in the convergence_test class in this module. Now you
perform the same tasks with

    import mg2
    HIST_FILE_NAME = "MG2_data_collection.cam.h1.0001-01-06-00000.nc"
    mgncol = 128
    ctest = mg2.convergence_test(HIST_FILE_NAME, mgncol)
    total_columns = 256
    timesteps = np.array([0.25, 1., 2., 5., 10., 15., 30., 60., 75., 100., 150., 300.])
    ctest.initialize(total_columns, timesteps)
    ctest.run_tests()
    ctest.plot_variable_convergence("QR", (100., 90., 50.))
    ctest.plot_column_variable_convergence("QR", (100., 90., 50.), 5.)